+++ cognite/client/_api/functions.py
@@ -70,7 +70,6 @@ def _get_function_identifier(function_id: Optional[int], function_external_id: O
 
 class FunctionsAPI(APIClient):
     _RESOURCE_PATH = "/functions"
-    _LIST_CLASS = FunctionList
 
     def __init__(self, config: ClientConfig, api_version: Optional[str], cognite_client: CogniteClient) -> None:
         super().__init__(config, api_version, cognite_client)
@@ -283,7 +282,7 @@ class FunctionsAPI(APIClient):
         ).dump(camel_case=True)
         res = self._post(url_path=f"{self._RESOURCE_PATH}/list", json={"filter": filter, "limit": limit})
 
-        return self._LIST_CLASS._load(res.json()["items"], cognite_client=self._cognite_client)
+        return FunctionList._load(res.json()["items"], cognite_client=self._cognite_client)
 
     def retrieve(
         self, id: Optional[int] = None, external_id: Optional[str] = None
@@ -658,7 +657,9 @@ def _sanitize_filename(filename: str) -> str:
 
 
 class FunctionCallsAPI(APIClient):
-    _LIST_CLASS = FunctionCallList
+    _RESOURCE_PATH = "/functions/{}/calls"
+    _RESOURCE_PATH_RESPONSE = "/functions/{}/calls/{}/response"
+    _RESOURCE_PATH_LOGS = "/functions/{}/calls/{}/logs"
 
     def list(
         self,
@@ -705,7 +706,7 @@ class FunctionCallsAPI(APIClient):
         filter = FunctionCallsFilter(
             status=status, schedule_id=schedule_id, start_time=start_time, end_time=end_time
         ).dump(camel_case=True)
-        resource_path = f"/functions/{function_id}/calls"
+        resource_path = self._RESOURCE_PATH.format(function_id)
         return self._list(
             method="POST",
             resource_path=resource_path,
@@ -747,7 +748,7 @@ class FunctionCallsAPI(APIClient):
         identifier = _get_function_identifier(function_id, function_external_id)
         function_id = _get_function_internal_id(self._cognite_client, identifier)
 
-        resource_path = f"/functions/{function_id}/calls"
+        resource_path = self._RESOURCE_PATH.format(function_id)
         identifiers = IdentifierSequence.load(ids=call_id).as_singleton()
 
         return self._retrieve_multiple(
@@ -759,7 +760,7 @@ class FunctionCallsAPI(APIClient):
 
     def get_response(
         self, call_id: int, function_id: Optional[int] = None, function_external_id: Optional[str] = None
-    ) -> Dict:
+    ) -> Optional[Dict]:
         """`Retrieve the response from a function call. <https://docs.cognite.com/api/v1/#operation/getFunctionCallResponse>`_
 
         Args:
@@ -768,7 +769,7 @@ class FunctionCallsAPI(APIClient):
             function_external_id (str, optional): External ID of the function on which the call was made.
 
         Returns:
-            Response from the function call.
+            Dict[str, Any] | None: Response from the function call.
 
         Examples:
 
@@ -788,9 +789,9 @@ class FunctionCallsAPI(APIClient):
         """
         identifier = _get_function_identifier(function_id, function_external_id)
         function_id = _get_function_internal_id(self._cognite_client, identifier)
-        url = f"/functions/{function_id}/calls/{call_id}/response"
-        res = self._get(url)
-        return res.json().get("response")
+
+        resource_path = self._RESOURCE_PATH_RESPONSE.format(function_id, call_id)
+        return self._get(resource_path).json().get("response")
 
     def get_logs(
         self, call_id: int, function_id: Optional[int] = None, function_external_id: Optional[str] = None
@@ -824,14 +825,12 @@ class FunctionCallsAPI(APIClient):
         identifier = _get_function_identifier(function_id, function_external_id)
         function_id = _get_function_internal_id(self._cognite_client, identifier)
 
-        url = f"/functions/{function_id}/calls/{call_id}/logs"
-        res = self._get(url)
-        return FunctionCallLog._load(res.json()["items"])
+        resource_path = self._RESOURCE_PATH_LOGS.format(function_id, call_id)
+        return FunctionCallLog._load(self._get(resource_path).json()["items"])
 
 
 class FunctionSchedulesAPI(APIClient):
     _RESOURCE_PATH = "/functions/schedules"
-    _LIST_CLASS = FunctionSchedulesList
 
     def __init__(self, config: ClientConfig, api_version: Optional[str], cognite_client: CogniteClient) -> None:
         super().__init__(config, api_version, cognite_client)
@@ -845,7 +844,6 @@ class FunctionSchedulesAPI(APIClient):
 
         Returns:
             Optional[FunctionSchedule]: Requested function schedule.
-
 
         Examples:
 
@@ -916,8 +914,9 @@ class FunctionSchedulesAPI(APIClient):
         ).dump(camel_case=True)
         res = self._post(url_path=f"{self._RESOURCE_PATH}/list", json={"filter": filter, "limit": limit})
 
-        return self._LIST_CLASS._load(res.json()["items"], cognite_client=self._cognite_client)
+        return FunctionSchedulesList._load(res.json()["items"], cognite_client=self._cognite_client)
 
+    # TODO: Major version 7, remove 'function_external_id' which only worked when using API-keys.
     def create(
         self,
         name: str,
@@ -933,41 +932,53 @@ class FunctionSchedulesAPI(APIClient):
         Args:
             name (str): Name of the schedule.
             function_id (optional, int): Id of the function. This is required if the schedule is created with client_credentials.
-            function_external_id (optional, str): External id of the function. This is deprecated and cannot be used together with client_credentials.
+            function_external_id (optional, str): External id of the function. **NOTE**: This is deprecated and will be removed in a future major version.
             description (str): Description of the schedule.
             cron_expression (str): Cron expression.
             client_credentials: (optional, ClientCredentials, Dict): Instance of ClientCredentials or a dictionary containing client credentials:
                 client_id
                 client_secret
-            data (optional, Dict): Data to be passed to the scheduled run. **WARNING:** Secrets or other confidential information should not be passed via this argument. There is a dedicated `secrets` argument in FunctionsAPI.create() for this purpose.
+            data (optional, Dict): Data to be passed to the scheduled run.
 
         Returns:
             FunctionSchedule: Created function schedule.
 
+        Warning:
+            Do not pass secrets or other confidential information via the ``data`` argument. There is a dedicated
+            ``secrets`` argument in FunctionsAPI.create() for this purpose.
+
         Examples:
 
-            Create function schedule::
+            Create a function schedule that runs using specified client credentials (**recommended**)::
 
+                >>> import os
                 >>> from cognite.client import CogniteClient
                 >>> from cognite.client.data_classes import ClientCredentials
                 >>> c = CogniteClient()
                 >>> schedule = c.functions.schedules.create(
-                ...     name= "My schedule",
+                ...     name="My schedule",
                 ...     function_id=123,
                 ...     cron_expression="*/5 * * * *",
-                ...     client_credentials=ClientCredentials("my-client-id", "my-client-secret"),
-                ...     description="This schedule does magic stuff."
+                ...     client_credentials=ClientCredentials("my-client-id", os.environ["MY_CLIENT_SECRET"]),
+                ...     description="This schedule does magic stuff.",
+                ...     data={"magic": "stuff"},
+                ... )
+
+            You may also create a schedule that runs with your -current- credentials, i.e. the same credentials you used
+            to instantiate the ``CogniteClient`` (that you're using right now). **Note**: Unless you happen to already use
+            client credentials, *this is not a recommended way to create schedules*, as it will create an explicit dependency
+            on your user account, which it will run the function "on behalf of" (until the schedule is eventually removed)::
+
+                >>> schedule = c.functions.schedules.create(
+                ...     name="My schedule",
+                ...     function_id=456,
+                ...     cron_expression="*/5 * * * *",
+                ...     description="A schedule just used for some temporary testing.",
                 ... )
 
         """
         _get_function_identifier(function_id, function_external_id)
-
-        nonce = None
-        if client_credentials is not None:
-            if function_id is None:
-                raise ValueError("When passing 'client_credentials', 'function_id' must be set")
-            nonce = _create_session_and_return_nonce(self._cognite_client, client_credentials)
-
+        nonce = _create_session_and_return_nonce(self._cognite_client, client_credentials)
         body: Dict[str, List[Dict[str, Union[str, int, None, Dict]]]] = {
             "items": [
                 {